+++ src/components/head.tsx
@@ -1,11 +1,12 @@
 import { useEffect } from "preact/hooks";
 import { useGlobalState } from "@ekwoka/preact-global-state";
 import { seti18Locale } from "../util/locale";
+import searchEngineData from "../assets/searchEngineData.json";
 
 function Head({ pageTitle }: HeadTypes) {
     const [theme, setTheme] = useGlobalState<string>("theme", localStorage.getItem("metallic/theme") || "default");
     const [service, setService] = useGlobalState<string>("service", localStorage.getItem("metallic/service") || "ultraviolet");
-    const [searchEngine, setSearchEngine] = useGlobalState<string>("engine", localStorage.getItem("metallic/engine") || "google");
+    const [searchEngine, setSearchEngine] = useGlobalState<string>("engine", localStorage.getItem("metallic/engine") || searchEngineData.google);
     const [title, setTitle] = useGlobalState<string>("title", localStorage.getItem("metallic/title") || "");
     const [icon, setIcon] = useGlobalState<string>("icon", localStorage.getItem("metallic/icon") || "");
     const [locale, setLocale] = useGlobalState<string>("locale", localStorage.getItem("metallic/locale") || "en");