+++ src/pages/settings/search.tsx
@@ -1,6 +1,7 @@
 import { useTranslation } from "react-i18next";
 import { useGlobalState } from "@ekwoka/preact-global-state";
 import { SquareButton } from "../../interface/button";
+import { SquareInput } from "../../interface/input";
 import { rammerhead } from "../../settings";
 import { Obfuscated } from "../../util/obfuscate";
 import ultravioletLogo from "../../assets/services/ultravioletLogo.png";
@@ -14,11 +15,12 @@ import { BraveLogo } from "../../assets/searchEngines/braveLogo";
 import { QwantLogo } from "../../assets/searchEngines/qwantLogo";
 import { SearXNGLogo } from "../../assets/searchEngines/searXNGLogo";
 import { EcosiaLogo } from "../../assets/searchEngines/ecosiaLogo";
+import searchEngineData from "../../assets/searchEngineData.json";
 
 function SearchSettings() {
     const { t } = useTranslation();
     const [service, setService] = useGlobalState<string>("service", localStorage.getItem("metallic/service") || "ultraviolet");
-    const [searchEngine, setSearchEngine] = useGlobalState<string>("engine", localStorage.getItem("metallic/engine") || "google");
+    const [searchEngine, setSearchEngine] = useGlobalState<string>("engine", localStorage.getItem("metallic/engine") || searchEngineData.google);
     const [openUsing, setOpenUsing] = useGlobalState<string>("open", localStorage.getItem("metallic/open") || "default");
 
     return (
@@ -42,38 +44,42 @@ function SearchSettings() {
             </section>
             <h1 class="text-4xl font-bold my-8">{t("settings.search.searchEngine.title")}</h1>
             <section class="flex flex-wrap items-center gap-3">
-                <SquareButton active={searchEngine == "google"} onClick={() => setSearchEngine("google")}>
+                <SquareButton active={searchEngine == searchEngineData.google} onClick={() => setSearchEngine(searchEngineData.google)}>
                     <GoogleLogo />
                     <span class="font-bold">Google</span>
                 </SquareButton>
-                <SquareButton active={searchEngine == "duckduckgo"} onClick={() => setSearchEngine("duckduckgo")}>
+                <SquareButton active={searchEngine == searchEngineData.duckduckgo} onClick={() => setSearchEngine(searchEngineData.duckduckgo)}>
                     <DuckDuckGoLogo />
                     <span class="font-bold">DuckDuckGo</span>
                 </SquareButton>
-                <SquareButton active={searchEngine == "bing"} onClick={() => setSearchEngine("bing")}>
+                <SquareButton active={searchEngine == searchEngineData.bing} onClick={() => setSearchEngine(searchEngineData.bing)}>
                     <BingLogo />
                     <span class="font-bold">Bing</span>
                 </SquareButton>
-                <SquareButton active={searchEngine == "yahoo"} onClick={() => setSearchEngine("yahoo")}>
+                <SquareButton active={searchEngine == searchEngineData.yahoo} onClick={() => setSearchEngine(searchEngineData.yahoo)}>
                     <YahooLogo />
                     <span class="font-bold">Yahoo</span>
                 </SquareButton>
-                <SquareButton active={searchEngine == "brave"} onClick={() => setSearchEngine("brave")}>
+                <SquareButton active={searchEngine == searchEngineData.brave} onClick={() => setSearchEngine(searchEngineData.brave)}>
                     <BraveLogo />
                     <span class="font-bold">Brave</span>
                 </SquareButton>
-                <SquareButton active={searchEngine == "qwant"} onClick={() => setSearchEngine("qwant")}>
+                <SquareButton active={searchEngine == searchEngineData.qwant} onClick={() => setSearchEngine(searchEngineData.qwant)}>
                     <QwantLogo />
                     <span class="font-bold">Qwant</span>
                 </SquareButton>
-                <SquareButton active={searchEngine == "searxng"} onClick={() => setSearchEngine("searxng")}>
+                <SquareButton active={searchEngine == searchEngineData.searxng} onClick={() => setSearchEngine(searchEngineData.searxng)}>
                     <SearXNGLogo />
                     <span class="font-bold">SearXNG</span>
                 </SquareButton>
-                <SquareButton active={searchEngine == "ecosia"} onClick={() => setSearchEngine("ecosia")}>
+                <SquareButton active={searchEngine == searchEngineData.ecosia} onClick={() => setSearchEngine(searchEngineData.ecosia)}>
                     <EcosiaLogo />
                     <span class="font-bold">Ecosia</span>
                 </SquareButton>
+            </section>
+            <h1 class="text-4xl font-bold my-8"><Obfuscated>{t("settings.search.customSearchEngine.title")}</Obfuscated></h1>
+            <section class="flex flex-wrap items-center gap-3">
+                <SquareInput placeholder={t("settings.search.customSearchEngine.input")} value={searchEngine} onInput={(e: any) => setSearchEngine(e.target.value)} />
             </section>
             <h1 class="text-4xl font-bold my-8">{t("settings.search.open.title")}</h1>
             <section class="flex flex-wrap items-center gap-3">